# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "swagger 接口文档",
        "contact": {},
        "version": "2.0"
    },
    "host": "localhost:9101",
    "paths": {
        "/ping": {
            "get": {
                "description": "连接测试",
                "summary": "ping",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/user/log/retrieve": {
            "get": {
                "summary": "获取用户日志",
                "parameters": [
                    {
                        "type": "string",
                        "description": "事件名称",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "查询开始时间 YYYY-MM-DD",
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "查询结束时间 YYYY-MM-DD",
                        "name": "end",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/user/log/upload": {
            "post": {
                "summary": "添加用户日志",
                "parameters": [
                    {
                        "description": "json",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PushJson"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "models.EventInfo": {
            "type": "object",
            "properties": {
                "act_time": {
                    "type": "string"
                },
                "append": {
                    "type": "string"
                }
            }
        },
        "models.PushJson": {
            "type": "object",
            "properties": {
                "device_id": {
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                },
                "user_events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.UserEvent"
                    }
                }
            }
        },
        "models.UserEvent": {
            "type": "object",
            "properties": {
                "event_infos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.EventInfo"
                    }
                },
                "event_name": {
                    "type": "string"
                }
            }
        }
    }
}